+++ zcash_primitives/src/transaction/builder.rs
@@ -550,7 +550,7 @@ mod tests {
         memo::MemoBytes,
         sapling::{Node, Rseed},
         transaction::components::{
-            amount::{Amount, DEFAULT_FEE},
+            amount::Amount,
             sapling::builder::{self as sapling_builder},
             transparent::builder::{self as transparent_builder},
         },
@@ -646,7 +646,7 @@ mod tests {
         builder
             .add_transparent_output(
                 &TransparentAddress::PublicKey([0; 20]),
-                Amount::from_u64(49000).unwrap(),
+                Amount::from_u64(40000).unwrap(),
             )
             .unwrap();
 
@@ -682,7 +682,7 @@ mod tests {
         builder
             .add_transparent_output(
                 &TransparentAddress::PublicKey([0; 20]),
-                Amount::from_u64(49000).unwrap(),
+                Amount::from_u64(40000).unwrap(),
             )
             .unwrap();
 
@@ -711,6 +711,8 @@ mod tests {
 
     #[test]
     fn fails_on_negative_change() {
+        use crate::transaction::fees::zip317::MINIMUM_FEE;
+
         let mut rng = OsRng;
 
         // Just use the master key as the ExtendedSpendingKey for this test
@@ -725,7 +727,7 @@ mod tests {
             let builder = Builder::new(TEST_NETWORK, tx_height);
             assert_eq!(
                 builder.mock_build(),
-                Err(Error::InsufficientFunds(DEFAULT_FEE))
+                Err(Error::InsufficientFunds(MINIMUM_FEE))
             );
         }
 
@@ -734,7 +736,7 @@ mod tests {
         let to = dfvk.default_address().1;
 
         // Fail if there is only a Sapling output
-        // 0.0005 z-ZEC out, 0.00001 t-ZEC fee
+        // 0.0005 z-ZEC out, 0.0001 t-ZEC fee
         {
             let mut builder = Builder::new(TEST_NETWORK, tx_height);
             builder
@@ -748,13 +750,13 @@ mod tests {
             assert_eq!(
                 builder.mock_build(),
                 Err(Error::InsufficientFunds(
-                    (Amount::from_i64(50000).unwrap() + DEFAULT_FEE).unwrap()
+                    (Amount::from_i64(50000).unwrap() + MINIMUM_FEE).unwrap()
                 ))
             );
         }
 
         // Fail if there is only a transparent output
-        // 0.0005 t-ZEC out, 0.00001 t-ZEC fee
+        // 0.0005 t-ZEC out, 0.0001 t-ZEC fee
         {
             let mut builder = Builder::new(TEST_NETWORK, tx_height);
             builder
@@ -766,19 +768,19 @@ mod tests {
             assert_eq!(
                 builder.mock_build(),
                 Err(Error::InsufficientFunds(
-                    (Amount::from_i64(50000).unwrap() + DEFAULT_FEE).unwrap()
+                    (Amount::from_i64(50000).unwrap() + MINIMUM_FEE).unwrap()
                 ))
             );
         }
 
-        let note1 = to.create_note(50999, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)));
+        let note1 = to.create_note(59999, Rseed::BeforeZip212(jubjub::Fr::random(&mut rng)));
         let cmu1 = Node::from_cmu(&note1.cmu());
         let mut tree = CommitmentTree::<Node, 32>::empty();
         tree.append(cmu1).unwrap();
         let mut witness1 = IncrementalWitness::from_tree(tree.clone());
 
         // Fail if there is insufficient input
-        // 0.0003 z-ZEC out, 0.0002 t-ZEC out, 0.00001 t-ZEC fee, 0.00050999 z-ZEC in
+        // 0.0003 z-ZEC out, 0.0002 t-ZEC out, 0.0001 t-ZEC fee, 0.00059999 z-ZEC in
         {
             let mut builder = Builder::new(TEST_NETWORK, tx_height);
             builder